Add a memory type and set/get methods
Because Baby stores LSB in the left-most position and we want to
display it as it would have appeared, we need to reverse the
bits in the memory before we use them, extract instructions, etc. To
do this, abstract the memory in a new type with GetWord, GetRaw and
SetWord methods. Get and Set Word methods return int32 in the format
we'd expect from modern machines and we do all further work on those
normal values. The methods take care to call bits.Reverse32
appropriately. GetRaw will return that memory value without the call
to reverse for ease of display.